+++ src/Pages/sangyeonkim/Main/Main.scss
@@ -118,6 +118,10 @@
           color: lightgray;
         }
 
+        ul {
+          list-style: none;
+        }
+
         .total-search {
           display: flex;
         }
